added Image Field Under Form2

+++ consumer/view.py
@@ -27,7 +27,10 @@ class DeleteConsumerAudit(Resource):
         if user.role != "audit" and user.permission != "consumer":
             return jsonify({"message" : "'error': 'Unauthorized access'"})        
         try:
-            audit_id = request.args.get('audit_id')            
+            audit_id = request.args.get('audit_id')
+            if audit_id is not None:
+                return jsonify({'message': 'Audit ID already exists'})
+        
             audit_document = AuditData.objects(audit_id=audit_id).first()
             if audit_document:
                 audit_document.delete()
@@ -37,86 +40,7 @@ class DeleteConsumerAudit(Resource):
         except Exception as e:
             print("Exception: ", e)
             return jsonify({"message": "Error: {}".format(str(e))})
-       
 
-class CreateConsumerAudit(Resource):
-    @jwt_required()
-    def post(self):
-        try:
-            user = User.objects.get(id=get_jwt_identity()['id'])
-        except DoesNotExist:
-            return unauthorized()
-        if user.role != "audit" and user.permission != "consumer":
-            return jsonify({"message" : "'error': 'Unauthorized access'"})        
-        try:
-            # Access form data
-            data = request.form
-            # Parse form1 and form2 from JSON strings
-            form1_data = json.loads(data.get('form1'))
-            form2_data = json.loads(data.get('form2'))
-            # Create Form1 object
-            form1 = Form1(
-                supervisor_contact=form1_data.get('supervisor_contact'),
-                tech_pt=form1_data.get('tech_pt'),
-                vehicle_number=form1_data.get('vehicle_number'),
-                tech_skills=form1_data.get('tech_skills'),
-                sr_manager=form1_data.get('sr_manager'),
-                tech_fullname=form1_data.get('tech_fullname'),
-                region=form1_data.get('region'),
-                vendor=form1_data.get('vendor'),
-                director=form1_data.get('director'),
-                sr_number=form1_data.get('sr_number'),
-                tech_ein=form1_data.get('tech_ein'),
-                team=form1_data.get('team'),
-                duty_manager=form1_data.get('duty_manager'),
-                supervisor=form1_data.get('supervisor'),
-                shortdescription=form1_data.get('shortdescription'),
-                tech_contact=form1_data.get('tech_contact')
-            )
-            # Create Form2 objects
-            form2_objects = []
-            for key, value in form2_data.items():
-                form2_objects.append(Form2(
-                    violation=value.get('violation'),
-                    remarks=value.get('remarks')
-                ))
-            # Create AuditData object
-            audit_data = AuditData(
-                username=data.get('username'),
-                status=data.get('status'),
-                name=data.get('name'),
-                audit_id=int(data.get('audit_id')),
-                lastmodified=data.get('lastmodified'),
-                supervisor_id=int(data.get('supervisor_id')),
-                expiryDate=data.get('expiryDate'),
-                auditDate=data.get('auditDate'),
-                remarks=data.get('remarks'),
-                form1=form1,
-                department=data.get('department'),
-                createdDate=data.get('createdDate'),
-                auditor_id=int(data.get('auditor_id')),
-                description=data.get('description')
-            )
-            # Save form2 objects
-            for i, form2_obj in enumerate(form2_objects):
-                setattr(audit_data, f'ceqv0{i+1}', form2_obj)
-            # Save image files
-            for image_key, file_data in request.files.items():
-                if image_key in ['ceqv01_image', 'ceqv02_image', 'ceqv03_image', 'ceqv04_image', 'ceqv05_image', 'ceqv06_image']:
-                    image_stream = io.BytesIO(file_data.read())
-                    setattr(audit_data, image_key, image_stream)
-                elif image_key in ['audit_signature', 'audited_staff_signature']:
-                    signature_stream = io.BytesIO(file_data.read())
-                    setattr(audit_data, image_key, signature_stream)
-
-            # Save the audit data to the database
-            audit_data.save()
-
-            return jsonify({'message': 'Audit created successfully'})
-
-        except Exception as e:
-            print("Exception: ", e)
-            return jsonify({'message': 'Error occurred while creating audit'})
 
 class GetConsumerAudit(Resource):
     @jwt_required()
@@ -128,31 +52,31 @@ class GetConsumerAudit(Resource):
         
         if user.role != "audit" and user.permission != "consumer":
             return jsonify({"message" : "'error': 'Unauthorized access'"})
-        
         audit_id = request.args.get('audit_id')      
         
         try:
             # Retrieve audit data by audit_id
             audit_data = AuditData.objects(audit_id=audit_id).first()
+            if audit_data is not None:
+                return jsonify({'message': 'Audit ID already exists'})
 
             if audit_data:
                 # Serialize audit_data to JSON
                 audit_json = json.loads(audit_data.to_json())
-
-                # Convert image fields to base64 encoded strings
-                for image_key in ['ceqv01_image', 'ceqv02_image', 'ceqv03_image', 'ceqv04_image', 'ceqv05_image', 'ceqv06_image']:
+                ceqv_list = ['ceqv01', 'ceqv02', 'ceqv03', 'ceqv04', 'ceqv05', 'ceqv06']
+                for image_key in ceqv_list:
                     if hasattr(audit_data, image_key):
-                        image_stream = getattr(audit_data, image_key)
-                        if image_stream:
-                            audit_json[image_key] = base64.b64encode(image_stream.read()).decode('utf-8')
-
+                        image_stream = getattr(audit_data[image_key], 'image')
+                        if image_stream:                                   
+                            audit_json[image_key]['image'] = base64.b64encode(image_stream.read()).decode('utf-8')
+                                    
                 # Convert audit_signature and audited_staff_signature to base64 encoded strings
                 for signature_key in ['audit_signature', 'audited_staff_signature']:
                     if hasattr(audit_data, signature_key):
                         signature_stream = getattr(audit_data, signature_key)
                         if signature_stream:
                             audit_json[signature_key] = base64.b64encode(signature_stream.read()).decode('utf-8')
-
+            
                 return jsonify(audit_json)
             else:
                 return jsonify({'message': 'Audit not found'})
@@ -171,7 +95,7 @@ class GetConsumerAuditList(Resource):
         
         if user.role != "audit" and user.permission != "consumer":
             return jsonify({"message" : "'error': 'Unauthorized access'"})        
-            
+        
         try:
             # Retrieve all audit data
             all_audit_data = AuditData.objects()
@@ -179,19 +103,18 @@ class GetConsumerAuditList(Resource):
             if all_audit_data:
                 # Initialize an empty list to store serialized audit data
                 audit_list = []
-
                 # Iterate through each audit and serialize it to JSON
                 for audit_data in all_audit_data:
                     audit_json = json.loads(audit_data.to_json())
-
-                    # Convert image fields to base64 encoded strings
-                    for image_key in ['ceqv01_image', 'ceqv02_image', 'ceqv03_image', 'ceqv04_image', 'ceqv05_image', 'ceqv06_image']:
+                    # Convert image fields to base64 encoded strings    
+                    audit_json = json.loads(audit_data.to_json())
+                    ceqv_list = ['ceqv01', 'ceqv02', 'ceqv03', 'ceqv04', 'ceqv05', 'ceqv06']
+                    for image_key in ceqv_list:
                         if hasattr(audit_data, image_key):
-                            image_stream = getattr(audit_data, image_key)
-                            if image_stream:
-                                audit_json[image_key] = base64.b64encode(image_stream.read()).decode('utf-8')
-
-                    # Convert audit_signature and audited_staff_signature to base64 encoded strings
+                            image_stream = getattr(audit_data[image_key], 'image')
+                            if image_stream:                                   
+                                audit_json[image_key]['image'] = base64.b64encode(image_stream.read()).decode('utf-8')            
+                   # Convert audit_signature and audited_staff_signature to base64 encoded strings
                     for signature_key in ['audit_signature', 'audited_staff_signature']:
                         if hasattr(audit_data, signature_key):
                             signature_stream = getattr(audit_data, signature_key)
@@ -200,7 +123,6 @@ class GetConsumerAuditList(Resource):
 
                     # Append the serialized audit data to the audit_list
                     audit_list.append(audit_json)
-
                 return jsonify(audit_list)
             else:
                 return jsonify({'message': 'No audits found'})
@@ -209,28 +131,25 @@ class GetConsumerAuditList(Resource):
             print("Exception: ", e)
             return jsonify({'message': 'Error occurred while retrieving audits'})
 
-
 class UpdateConsumerAudit(Resource):
     @jwt_required()
     def post(self):
-        
         try:
             user = User.objects.get(id=get_jwt_identity()['id'])
         except DoesNotExist:
-            return unauthorized()  
-        
+            return unauthorized()       
         if (user.role == "audit" or user.role == "superviser" )  and (user.permission != "consumer"):
-            return jsonify({"message" : "'error': 'Unauthorized access'"})        
-                 
+            return jsonify({"message" : "'error': 'Unauthorized access'"})
         audit_id = request.args.get('audit_id')      
         try:
             # Retrieve the audit data by audit_id
             audit_data = AuditData.objects(audit_id=audit_id).first()
-
+            if audit_data.first() is not None:
+                return jsonify({'message': 'Audit ID already exists'})
+            
             if audit_data:
                 # Access form data
                 data = request.form
-
                 # Update Form1 object
                 form1_data = json.loads(data.get('form1'))
                 form1 = Form1(
@@ -260,7 +179,7 @@ class UpdateConsumerAudit(Resource):
                     if form2_obj:
                         form2_obj.violation = value.get('violation')
                         form2_obj.remarks = value.get('remarks')
-
+                
                 # Update other audit data fields
                 audit_data.username = data.get('username')
                 audit_data.status = data.get('status')
@@ -274,15 +193,15 @@ class UpdateConsumerAudit(Resource):
                 audit_data.createdDate = data.get('createdDate')
                 audit_data.auditor_id = int(data.get('auditor_id'))
                 audit_data.description = data.get('description')
-
                 # Update images
                 for image_key, file_data in request.files.items():
-                    if image_key in ['ceqv01_image', 'ceqv02_image', 'ceqv03_image', 'ceqv04_image', 'ceqv05_image', 'ceqv06_image']:
+                    if image_key.startswith('ceqv') and image_key.endswith('_image'):
                         image_stream = io.BytesIO(file_data.read())
-                        setattr(audit_data, image_key, image_stream)
+                        print(image_key)
+                        setattr(audit_data[image_key[0:6]], 'image', image_stream)  
                     elif image_key in ['audit_signature', 'audited_staff_signature']:
                         signature_stream = io.BytesIO(file_data.read())
-                        setattr(audit_data, image_key, signature_stream)
+                        setattr(audit_data, image_key, signature_stream) 
 
                 # Save the updated audit data
                 audit_data.save()
@@ -294,3 +213,74 @@ class UpdateConsumerAudit(Resource):
         except Exception as e:
             print("Exception: ", e)
             return jsonify({'message': 'Error occurred while updating audit'})
+
+class CreateConsumerAudit(Resource):
+    @jwt_required()
+    def post(self):
+        try:
+            user = User.objects.get(id=get_jwt_identity()['id'])
+        except DoesNotExist:
+            return unauthorized()
+        if user.role != "audit" and user.permission != "consumer":
+            return jsonify({"message": "Unauthorized access"}), 403
+
+        try:
+            # Access form data
+            data = request.form
+            form1_data = data.get('form1')
+            form2_data = data.get('form2')
+
+            # Convert JSON strings to Python dictionaries
+            form1_data = json.loads(form1_data)
+            form2_data = json.loads(form2_data)
+
+            # Create Form1 object
+            form1 = Form1(**form1_data)
+
+            # Create Form2 objects
+            form2_objects = []
+            for key, value in form2_data.items():
+                form2_obj = Form2(
+                    violation=value.get('violation'),
+                    remarks=value.get('remarks')
+                )
+                form2_objects.append(form2_obj)
+
+            # Create AuditData object
+            audit_data = AuditData(
+                username=data.get('username'),
+                status=data.get('status'),
+                name=data.get('name'),
+                audit_id=int(data.get('audit_id')),
+                lastmodified=data.get('lastmodified'),
+                supervisor_id=int(data.get('supervisor_id')),
+                expiryDate=data.get('expiryDate'),
+                auditDate=data.get('auditDate'),
+                remarks=data.get('remarks'),
+                form1=form1,
+                department=data.get('department'),
+                createdDate=data.get('createdDate'),
+                auditor_id=int(data.get('auditor_id')),
+                description=data.get('description')
+            )
+
+            # Save Form2 objects
+            for i, form2_obj in enumerate(form2_objects, start=1):
+                setattr(audit_data, f'ceqv{i:02}', form2_obj)
+
+            # Save image files
+            for image_key, file_data in request.files.items():
+                if image_key.startswith('ceqv') and image_key.endswith('_image'):
+                    image_stream = io.BytesIO(file_data.read())
+                    setattr(audit_data[image_key[0:6]], 'image', image_stream)  
+                elif image_key in ['audit_signature', 'audited_staff_signature']:
+                    signature_stream = io.BytesIO(file_data.read())
+                    setattr(audit_data, image_key, signature_stream)    
+            # Save the audit data to the database
+            audit_data.save()
+
+            return jsonify({'message': 'Audit created successfully'})
+
+        except Exception as e:
+            print("Exception:", e)
+            return jsonify({'message': 'Error occurred while creating audit'})
